+++ src/SMAPI.Mods.ConsoleCommands/Framework/Commands/TrainerCommand.cs
@@ -78,8 +78,7 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework.Commands
         /// <param name="data">The data to display.</param>
         /// <param name="header">The table header.</param>
         /// <param name="getRow">Returns a set of fields for a data value.</param>
-        /// <param name="rightAlign">Whether to right-align the data.</param>
-        protected string GetTableString<T>(IEnumerable<T> data, string[] header, Func<T, string[]> getRow, bool rightAlign = false)
+        protected string GetTableString<T>(IEnumerable<T> data, string[] header, Func<T, string[]> getRow)
         {
             // get table data
             int[] widths = header.Select(p => p.Length).ToArray();
@@ -108,7 +107,7 @@ namespace StardewModdingAPI.Mods.ConsoleCommands.Framework.Commands
             return string.Join(
                 Environment.NewLine,
                 lines.Select(line => string.Join(" | ",
-                    line.Select((field, i) => rightAlign ? field.PadRight(widths[i], ' ') : field.PadLeft(widths[i], ' '))
+                    line.Select((field, i) => field.PadLeft(widths[i], ' '))
                 ))
             );
         }