+++ src/SMAPI/Framework/CommandManager.cs
@@ -15,10 +15,20 @@ namespace StardewModdingAPI.Framework
         /// <summary>The commands registered with SMAPI.</summary>
         private readonly IDictionary<string, Command> Commands = new Dictionary<string, Command>(StringComparer.OrdinalIgnoreCase);
 
+        /// <summary>Writes messages to the console.</summary>
+        private readonly IMonitor Monitor;
+
 
         /*********
         ** Public methods
         *********/
+        /// <summary>Construct an instance.</summary>
+        /// <param name="monitor">Writes messages to the console.</param>
+        public CommandManager(IMonitor monitor)
+        {
+            this.Monitor = monitor;
+        }
+
         /// <summary>Add a console command.</summary>
         /// <param name="mod">The mod adding the command (or <c>null</c> for a SMAPI command).</param>
         /// <param name="name">The command name, which the user must type to trigger it.</param>
@@ -81,8 +91,9 @@ namespace StardewModdingAPI.Framework
         /// <param name="name">The parsed command name.</param>
         /// <param name="args">The parsed command arguments.</param>
         /// <param name="command">The command which can handle the input.</param>
+        /// <param name="screenId">The screen ID on which to run the command.</param>
         /// <returns>Returns true if the input was successfully parsed and matched to a command; else false.</returns>
-        public bool TryParse(string input, out string name, out string[] args, out Command command)
+        public bool TryParse(string input, out string name, out string[] args, out Command command, out int screenId)
         {
             // ignore if blank
             if (string.IsNullOrWhiteSpace(input))
@@ -90,6 +101,7 @@ namespace StardewModdingAPI.Framework
                 name = null;
                 args = null;
                 command = null;
+                screenId = 0;
                 return false;
             }
 
@@ -97,6 +109,27 @@ namespace StardewModdingAPI.Framework
             args = this.ParseArgs(input);
             name = this.GetNormalizedName(args[0]);
             args = args.Skip(1).ToArray();
+
+            // get screen ID argument
+            screenId = 0;
+            for (int i = 0; i < args.Length; i++)
+            {
+                // consume arg & set screen ID
+                if (this.TryParseScreenId(args[i], out int rawScreenId, out string error))
+                {
+                    args = args.Take(i).Concat(args.Skip(i + 1)).ToArray();
+                    screenId = rawScreenId;
+                    continue;
+                }
+
+                // invalid screen arg
+                if (error != null)
+                {
+                    this.Monitor.Log(error, LogLevel.Error);
+                    command = null;
+                    return false;
+                }
+            }
 
             // get command
             return this.Commands.TryGetValue(name, out command);
@@ -150,6 +183,38 @@ namespace StardewModdingAPI.Framework
             args.Add(currentArg.ToString());
 
             return args.Where(item => !string.IsNullOrWhiteSpace(item)).ToArray();
+        }
+
+        /// <summary>Try to parse a 'screen=X' command argument, which specifies the screen that should receive the command.</summary>
+        /// <param name="arg">The raw argument to parse.</param>
+        /// <param name="screen">The parsed screen ID, if any.</param>
+        /// <param name="error">The error which indicates an invalid screen ID, if applicable.</param>
+        /// <returns>Returns whether the screen ID was parsed successfully.</returns>
+        private bool TryParseScreenId(string arg, out int screen, out string error)
+        {
+            screen = -1;
+            error = null;
+
+            // skip non-screen arg
+            if (!arg.StartsWith("screen="))
+                return false;
+
+            // get screen ID
+            string rawScreen = arg.Substring("screen=".Length);
+            if (!int.TryParse(rawScreen, out screen))
+            {
+                error = $"invalid screen ID format: {rawScreen}";
+                return false;
+            }
+
+            // validate ID
+            if (!Context.HasScreenId(screen))
+            {
+                error = $"there's no active screen with ID {screen}. Active screen IDs: {string.Join(", ", Context.ActiveScreenIds)}.";
+                return false;
+            }
+
+            return true;
         }
 
         /// <summary>Get a normalized command name.</summary>