+++ docs/release-notes.md
@@ -7,6 +7,21 @@
   * Migrated to Harmony 2.0 (see [_migrate to Harmony 2.0_](https://stardewvalleywiki.com/Modding:Migrate_to_Harmony_2.0) for more info).
 -->
 
+## 3.8.3
+Released 08 January 2021 for Stardew Valley 1.5.2 or later.
+
+* For players:
+  * Updated for Stardew Valley 1.5.2.
+  * Reduced memory usage.
+  * You can now enter console commands for a specific screen in split-screen mode by adding `screen=ID` to the command.
+  * Typing `help` in the SMAPI console is now more helpful.
+
+* For modders:
+  * Simplified tilesheet order warning added in SMAPI 3.8.2.
+
+* For the Console Commands mod:
+  * Removed experimental `performance` command. Unfortunately this impacted SMAPI's memory usage and performance, and the data was often misinterpreted. This may be replaced with more automatic performance alerts in a future version.
+
 ## 3.8.2
 Released 03 January 2021 for Stardew Valley 1.5.1 or later.